Ambiguous description caused customer avoid using it
Ambiguous description caused the customer to avoid using it, we need to clarify actually what happens in this scenario a bit further. this transformer actually decodes the encoded input String and later Dataweave parse/transform it into the type defined in output header of script. For example, is DW header is output application/json, then it will transform the decoded string to application/json