Update scheduler-xml-reference.adoc
- corrected default value of "disallowConcurrentExecution" attribute, set to "false"
- corrected spelling of "disallowConcurrentExecution" attribute

rationale:
After seeing some unexpected behavior in our applications and running some tests, if the "disallowConcurrentExecution" attribute is not added then concurrent executions of the flow occurs with a scheduler endpoint.  So the default is "false" and not "true".
Tested by adding a transform with a "wait(10000)" to a flow with a scheduler endpoint set to run on fixed frequency of 1000ms.  So scheduler executes every 1 second, and the transform within it waits 10 seconds.  
Without setting  "disallowConcurrentExecution", the runtime executes the flow every 1 second, resulting in 10 concurrent executions.
Setting  "disallowConcurrentExecution" to "true", the runtime executes the flow every 10 seconds, resulting in 0 concurrent executions.