Update scheduler-concept.adoc
The statement needs to be updated. After the Mule runtime 4.1  default value for disallowConcurrentExecution is set to false which allows the next flow execution in the middle of the first execution in progress.  In the Mule runtime 4.1 and before default value for disallowConcurrentExecution is set to true hence if a flow is executing, the flow won’t trigger again until it finishes processing.

Reference Documentation :  
mule 4.1 https://docs.mulesoft.com/mule-runtime/4.1/scheduler-xml-reference#scheduler-properties
mule 4.4 https://docs.mulesoft.com/mule-runtime/4.4/scheduler-xml-reference#scheduler-properties