Removed the synchronized="false"
Removed the synchronized="false" in the line number 75 as per the line 63 this setting is for disabling the cache synchronization and line 70 says XML snippet is for the configuration of a caching strategy that synchronizes access to the cache. But with the setting synchronized="false" I believe we are actually disabling synchronization and that is not the intent as per line 70.