Update global-settings-configuration.adoc
"so the default value of 5000 milliseconds specifies that Mule has ten seconds to process and dispatch messages gracefully after shutdown is initiated."

It is not clear to me how a default value of 5000 milliseconds results in ten seconds of processing time.  Should this have been five seconds of processing time?